fix(init)

+++ init.sh
@@ -26,8 +26,16 @@ ln -nfs "${DOTFILES_ROOT}/.config/sheldon" "${HOME}/.config/"
 ln -nfs "${DOTFILES_ROOT}/.config/wezterm" "${HOME}/.config/"
 
 mkdir -p "${HOME}/.config/karabiner/assets"
+if [[ -e "${HOME}/.config/karabiner/karabiner.json" ]]; then
+  echo "Error: ${HOME}/.config/karabiner/karabiner.json already exists"
+  exit 1
+fi
 ln -nfs "${DOTFILES_ROOT}/.config/karabiner/karabiner.json" "${HOME}/.config/karabiner/karabiner.json"
-ln -nfs "${DOTFILES_ROOT}/.config/karabiner/assets/complex_modifications/" "${HOME}/.config/karabiner/assets/complex_modifications/"
+if [[ -e "${HOME}/.config/karabiner/assets/complex_modifications" ]]; then
+  echo "Error: ${HOME}/.config/karabiner/assets/complex_modifications already exists"
+  exit 1
+fi
+ln -nfs "${DOTFILES_ROOT}/.config/karabiner/assets/complex_modifications" "${HOME}/.config/karabiner/assets/complex_modifications"
 
 mkdir -p "${HOME}/.config/git/"
 touch "${HOME}/.config/git/.gitconfig.local"
